+++ src/termux_tasker/ui/screens/install_task_version.py
@@ -28,7 +28,7 @@ from termux_tasker.ui.screens._utils import (
     git_checkout,
     get_installed_task_version,
     make_unique,
-    merge_runner_properties,
+    merge_task_properties,
     fill_default_properties,
     sanitize_id,
     is_property_value_empty,
@@ -186,7 +186,7 @@ class InstallTaskVersionScreen(MenuScreen):
         else:
             old_settings = TaskSettings.load(target_dir / "settings.toml")
             old_meta = TaskMetadata.load(target_dir / "metadata.toml")
-            new_settings = merge_runner_properties(
+            new_settings = merge_task_properties(
                 old_settings, old_meta.properties, meta.properties
             )
             new_settings.save(self.tmp_task_folder / "settings.toml")
@@ -289,7 +289,7 @@ class InstallTaskVersionScreen(MenuScreen):
         from termux_tasker.ui.screens.task_menu import TaskMenuScreen
 
         fill_default_properties(
-            self.tmp_task_folder / "settings.toml", meta.properties
+            self.tmp_task_folder / "settings.toml", meta.properties, kind="task"
         )
 
         if target_dir.exists():